Complete renaming in PagePatterns
