+++ packages/edit-site/src/components/page-main/index.js
@@ -6,7 +6,7 @@ import { privateApis as routerPrivateApis } from '@wordpress/router';
 /**
  * Internal dependencies
  */
-import PageLibrary from '../page-library';
+import PagePatterns from '../page-patterns';
 import PageTemplateParts from '../page-template-parts';
 import PageTemplates from '../page-templates';
 import { unlock } from '../../lock-unlock';
@@ -23,7 +23,7 @@ export default function PageMain() {
 	} else if ( path === '/wp_template_part/all' ) {
 		return <PageTemplateParts />;
 	} else if ( path === '/patterns' ) {
-		return <PageLibrary />;
+		return <PagePatterns />;
 	}
 
 	return null;
